requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
filetype==1.2.0
forbiddenfruit==0.1.4
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-core==0.3.61
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-api==0.2.34
langgraph-checkpoint==2.0.26
langgraph-cli==0.2.10
langgraph-prebuilt==0.2.0
langgraph-runtime-inmem==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
truststore==0.10.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
xxhash==3.5.0
zstandard==0.23.0
